WayCost#calculate_distance: Use Geokit instead of RGeo
Since the `WayCost` coordinates are already stored in Geokit objects,
use it to calculate distance instead of RGeo. This simplifies things by
allowing us to avoid type conversion.

I had previously used RGeo because my reading of the Geokit README
indicated that its purpose was only for storing and representing
geographic data types, not for making calculations on them. However,
Alban pointed out that this isn't the case, and it in fact does have a
distance method.

Needed a little extra manipulation to convert the value into the desired
metres unit as Geokit only supports kilometres, but otherwise it cuts
out a couple lines of code.

Refs #6952