CleanUp: Add `#run_methods`
This method is coupled with a new virtual attribute that can be used in
the initializer like:

    CleanUp.new(methods: [:destroy_routes])

The method will run all methods specified in the `:methods` list. The
plan is to replace the calls to `destroy_routes` etc. with a call to
this method. The result will be a more configurable clean-up process,
allowing users to selectively choose what to clean up by declaring what
methods in the `CleanUp` model to call.

Refs #6854