CleanUp: Add `#destroy_empty`
Gives us a single call site to trigger the destruction of vehicle
journeys, journey patterns, and routes without content.

This was previously done directly in the `#clean` method, but since it's
not needed during referential duplication (only during merges), we don't
want to enable it by default. Thus now to activate the same old
functionality, you would create a `CleanUp` like this:

    CleanUp.new(methods: [:destroy_empty])

Refs #6854